% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{codebook.ftp}
\alias{codebook.ftp}
\title{Ejecución de CodeBook}
\usage{
codebook.ftp()
}
\value{
Generates "geoftps.rds" RDS file
}
\description{
Función que inicializa una serie de valirables para simular el comportamiento definido en
el codigo facilitado en la práctica.
}
\details{
Dato unas variables llamamos a lo que sería el equivalente a nuestro main, el corzando del progrma
}
\examples{
codebook.ftp()
}
