% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{codebook}
\alias{codebook}
\title{Funcion general}
\usage{
codebook(url, sample, output.name)
}
\arguments{
\item{sample}{Numero de muestras.}

\item{output.name}{Nombre del fichero de salida donde guardaremos los datos.}
}
\value{
Generates output.name RDS file
}
\description{
Función que ejecuta ordenadamente todas las funciones decladaras, y simula el comportamiento
del codigo facilitado en la práctica.
}
\examples{
codebook(sample, output.name)
}
