% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{create.dir.data}
\alias{create.dir.data}
\title{Title}
\usage{
create.dir.data()
}
\description{
Title
}
\details{
Download file from 'download.url' and save it as 'destination.file' in dir.data
Path is relative to working directory
}
