% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{download.data}
\alias{download.data}
\title{Download file}
\usage{
download.data(url, destfile)
}
\arguments{
\item{destination.file}{}
}
\description{
Download file
}
\details{
Download file from 'download.url' and save it as 'destination.file' in dir.data
Path is relative to working directory
}
