% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{get.cve}
\alias{get.cve}
\title{Return data frame of CVEs for a specific CPE}
\usage{
get.cve(cpe, cves = NULL)
}
\arguments{
\item{cpe}{CPE}

\item{cves}{dataframe with cves}
}
\value{
score
}
\description{
Return data frame of CVEs for a specific CPE
}
