% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{get.opendata}
\alias{get.opendata}
\title{Get opendata csv or json file from rapid7.com and convert it to dataframe object}
\usage{
get.opendata(opendata.url = NULL, verbose)
}
\arguments{
\item{opendata.url}{Any url of https://opendata.rapid7.com}

\item{verbose}{Print logs}
}
\value{
df.tcp21
}
\description{
Función que se conecta a la web de rapid7 que lleva implicita, y se descarga un fichero en formato gz, el cual posteriormente lo descomprime y nos lo retorna en formato csv, borrando el fichero comprimido.
}
\details{
Funcion que nos retorna un fichero con las IP que tienen abierto el puerto TCP
}
\examples{
get.opendata(url, verbose)
}
