% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{get.score.max}
\alias{get.score.max}
\title{Return maxim score of cvss2 or cvss3}
\usage{
get.score.max(df.cve, cvss = 2)
}
\arguments{
\item{df.cve}{data frame with list of CVEs}

\item{cvss}{CVSS version. 2 or 3}
}
\value{
max.score
}
\description{
Return maxim score of cvss2 or cvss3
}
