% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{get.score}
\alias{get.score}
\title{Return cpe22 or cpe23 from vendor and version}
\usage{
get.score(df.cve, cvss = 2)
}
\arguments{
\item{df.cve}{data frame with list of CVEs}

\item{cvss}{CVSS version. 2 or 3}
}
\value{
score
}
\description{
Return cpe22 or cpe23 from vendor and version
}
