% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{http.responses}
\alias{http.responses}
\title{Generaci칩n del fichero de muestras}
\usage{
http.responses(port = 80, rows = 50, verbose = TRUE)
}
\arguments{
\item{port}{Nos indica el puerto del que estamos realizando el analisis}

\item{rows}{Filas que queremos extraer}

\item{verbose}{Si queremos mostrar por consola m치s informaci칩n de las operaciones que va realizando}
}
\value{
df Dataframe con las 50 muestras
}
\description{
Funci칩n que utiliza el fichero usado en el informe Rmd
}
