% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{latlong2map}
\alias{latlong2map}
\title{Generación de mapa por pais}
\usage{
latlong2map(pointsDF, mapping)
}
\arguments{
\item{pointsDF}{Dataframe donde donde tenemos la geolocalización}

\item{mapping}{Tipo de mapa que queremos pintar "world"}
}
\description{
Funcion que te calcula el pais dada la long/lat
}
