% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{syso}
\alias{syso}
\title{Imprirmir por consola}
\usage{
syso(x)
}
\arguments{
\item{x}{Texto que queremos imprimir}
}
\description{
Funcion que imprime por pantalla, simpre y cuando este habilitada a true la variable verbose
}
\details{
No se le pasa ningun valor a la función ya que dejamos como predefinido "data" por defecto, sin capacidad de ser alterado.
}
\examples{
syso("Hello World!!")
}
