% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{codebook}
\alias{codebook}
\title{Funcion general}
\usage{
codebook(url, sample, output.name, verbose = FALSE)
}
\arguments{
\item{url}{Url de opendata.rapid7.com con el fichero a evaluar}

\item{sample}{Numero de muestras.}

\item{output.name}{Nombre del fichero de salida donde guardaremos los datos.}

\item{verbose}{Mostrar comentarios}
}
\value{
Generates output.name RDS file
}
\description{
Función que ejecuta ordenadamente todas las funciones decladaras, y simula el comportamiento
del codigo facilitado en la práctica.
}
\examples{
codebook("https://opendata.rapid7.com/sonar.tcp/2019-04-21-1555815947-http_get_5555.csv.gz", 500, "output.rds", TRUE)
}
