% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{geolocate.http.responses}
\alias{geolocate.http.responses}
\title{Geolocate ip address using maxmind}
\usage{
geolocate.http.responses(df.maxmind, df.raw)
}
\arguments{
\item{df.raw}{}
}
\value{
df.raw.geo
}
\description{
Used in UntrustedServers.Rmd
}
