% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{get.cpe}
\alias{get.cpe}
\title{Return cpe22 or cpe23 from vendor and version}
\usage{
get.cpe(vendor, version, cpe.version = 22, cpes = NULL)
}
\arguments{
\item{cpes}{}
}
\value{
cpe
}
\description{
Requiered net.security package updated.
}
