% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{get.opendata}
\alias{get.opendata}
\title{Get opendata csv or json file from rapid7.com and convert it into dataframe object.}
\usage{
get.opendata(opendata.url = NULL, verbose)
}
\arguments{
\item{opendata.url}{Any url of https://opendata.rapid7.com}

\item{verbose}{Print logs}
}
\value{
dataframe
}
\description{
This function downloads a compressed '.gz' file from opendata.rapid7.com website, and imports file from csv or json into dataframe object.
}
\examples{
get.opendata(url, verbose)
}
