% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{parse.headers}
\alias{parse.headers}
\title{Generación del fichero de muestras}
\usage{
parse.headers(df.raw)
}
\arguments{
\item{df.raw}{}
}
\value{
df.raw dataframe con los datos limpios
}
\description{
Función que genera el fichero que utlizaremos para el informe, para ello
}
\examples{
parse.headers(port=80, rows=50, verbose=TRUE)
}
