% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{parse.headers}
\alias{parse.headers}
\title{Parsing headers of dataframe with data column encoded in base64}
\usage{
parse.headers(df.raw)
}
\value{
parsed df
}
\description{
Decode data in base64, extract headers, status, http version and server
}
\examples{
parse.headers(df.raw)
}
