% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{parse.headers}
\alias{parse.headers}
\title{Generaci칩n del fichero de muestras}
\usage{
parse.headers(port = 80, rows = 50, verbose = TRUE)
}
\arguments{
\item{port}{Nos indica el puerto del que estamos realizando el analisis}

\item{rows}{Filas que queremos extraer}

\item{verbose}{Si queremos mostrar por consola m치s informaci칩n de las operaciones que va realizando}
}
\value{
df.raw dataframe con los datos limpios
}
\description{
Funci칩n que genera el fichero que utlizaremos para el informe, para ello
}
\examples{
parse.headers(port=80, rows=50, verbose=TRUE)
}
